make unstick code more generic (buggy)
When bot is considered stuck, if bot is close enough to a friendly goal
(target), then reset stuck time.
This allows to not trigger unstick while healing or repairing.
OTOH, this genericity brings new problems and does not fix old ones:

Long unstick delays are still required when human bots target alien
structures, since they stand still while doing so. This is not much of
a problem, since unstick code simply makes them weirder to hit, but I
believe the correct fix here would be to make bots move while attacking
buildings;

Bots can still be stuck if ordered to not destroy buildings because
they never change goal;

Bots can now be stuck indefinitely if close enough to their target.
The cases where this can happen are, when a human bots try to repair
and ended up on the building, somehow, or if alien's crazy jumping
put them close enough to their heal source.
Here, previous code, which was not handling self-healing case, was
only built to reset when repairing, which resetted stuck time when
target was not healed and bot was using a CKit, which imply it was
not subject to this problem, since target would end up fully healed
at some point.

I think those points should be addressed before merging.